% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\docType{class}
\name{AmericanOption-class}
\alias{AmericanOption}
\alias{AmericanOption-class}
\title{A standard option contract allowing for \emph{early} exercise at the choice of the option holder}
\description{
A standard option contract allowing for \emph{early} exercise at the choice of the option holder
}
\section{Methods}{

\describe{
\item{\code{optionality_fcn(v, ...)}}{Return a version of {v} at time {t} corrected for any optionality conditions.}

\item{\code{recovery_fcn(v, S, t, ...)}}{Return recovery value, given non-default values {v} at time {t}.  Subclasses may be more elaborate, this method simply returns 0.0.}
}}

