% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackscholes.R
\docType{data}
\name{CALL}
\alias{CALL}
\title{Constant CALL for defining option contracts}
\format{An object of class \code{numeric} of length 1.}
\usage{
CALL
}
\description{
Constant CALL for defining option contracts
}
\keyword{datasets}
