% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/american_options.R
\name{control_variate_pairs}
\alias{control_variate_pairs}
\title{Form instrument objects for vanilla options}
\usage{
control_variate_pairs(callput, K, time)
}
\arguments{
\item{callput}{1 for calls, -1 for puts}

\item{K}{strike}

\item{time}{Time from \code{0} until expiration}
}
\description{
Form a list twice as long as the longest of the arguments
\code{callput}, \code{K}, \code{time} whose
first half consists of AmericanOption objects and
second half consists of EuropeanOption objects
having the same exercise specification
}
\seealso{
Other American Exercise Equity Options: 
\code{\link{american_implied_volatility}()},
\code{\link{american}()}
}
\concept{American Exercise Equity Options}
