% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\docType{class}
\name{EquityOption-class}
\alias{EquityOption}
\alias{EquityOption-class}
\title{An option contract with call or put terms}
\description{
An option contract with call or put terms
}
\section{Fields}{

\describe{
\item{\code{strike}}{A decision price for the contract}

\item{\code{callput}}{Either \code{1} for a call or \code{-1} for a put}
}}

