% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeks.R
\name{find_greeks}
\alias{find_greeks}
\title{Find present value and selected greeks for given set of instruments}
\usage{
find_greeks(
  S0,
  num_time_steps,
  instruments,
  greeks = c("delta", "gamma"),
  const_volatility = 0.5,
  const_short_rate = 0,
  const_default_intensity = 0,
  discount_factor_fcn = function(T, t, ...) {
     exp(-const_short_rate * (T - t))
 },
  default_intensity_fcn = function(t, S, ...) {
     const_default_intensity + 0 * S
 },
  variance_cumulation_fcn = function(T, t) {
     const_volatility^2 * (T - t)
 },
  dvola = NULL,
  dr = NULL,
  dhazard = NULL,
  ...
)
}
\arguments{
\item{S0}{An initial stock price, for setting grid scale}

\item{num_time_steps}{Minimum number of time steps in the grid}

\item{instruments}{A list of instruments to be priced.  Each
one must have a \code{strike} and a \code{optionality_fcn}, as
with \code{\link{GridPricedInstrument}} and its subclasses.}

\item{greeks}{Sequence of names of greeks to be estimated (see
\code{\link{GREEK_NAMES}} for the available values).  Delta and Gamma are
essentially zero-cost to include since they are read directly off the pricing
grid as spline derivatives.  All other greeks require repeated calls to the
pricing engine and multiply computation expense accordingly.}

\item{const_volatility}{A constant to use for volatility in case \code{variance_cumulation_fcn}
is not given}

\item{const_short_rate}{A constant to use for the instantaneous interest rate in case \code{discount_factor_fcn}
is not given}

\item{const_default_intensity}{A constant to use for the instantaneous default intensity in case \code{default_intensity_fcn}
is not given}

\item{discount_factor_fcn}{A function for computing present values to
time \code{t} of various cashflows occurring during this timestep, with
arguments \code{T}, \code{t}}

\item{default_intensity_fcn}{A function for computing default intensity
occurring during this timestep, dependent on time and stock price, with
arguments \code{t}, \code{S}.}

\item{variance_cumulation_fcn}{A function for computing total stock variance
occurring during this timestep, with arguments \code{T}, \code{t}.  E.g. with
a constant volatility \eqn{s} this takes the form \eqn{(T-t)s^2}.}

\item{dvola}{User-specified change in volatility to use for Vega estimation by
multiple pricing calls.  Usually left NULL, for automatic selection.}

\item{dr}{User-specified change in risk-free rates to use for Rates DV01 estimation
by multiple pricing calls.  Usually left NULL, for automatic selection.}

\item{dhazard}{User-specified change in default intensity to use for Credit DV01
estimation by multiple pricing calls.  Usually left NULL, for automatic selection.}

\item{...}{
  Arguments passed on to \code{\link{find_present_value}}
  \describe{
    \item{\code{grid_center}}{A reasonable central value for the grid, defaults to S0 or an instrument strike}
    \item{\code{borrow_cost}}{Stock borrow cost, affecting the drift rate}
    \item{\code{dividend_rate}}{Continuous dividend rate, affecting the drift rate}
    \item{\code{override_Tmax}}{A different maximum time on the grid to enforce}
    \item{\code{structure_constant}}{The maximum ratio between time intervals \code{dt}
and the square of space intervals \code{dz^2}}
    \item{\code{std_devs_width}}{The number of standard deviations, in \code{sigma * sqrt(T)}
units, to incorporate into the grid}
    \item{\code{dividends}}{A \code{data.frame} with columns \code{time}, \code{fixed},
and \code{proportional}.  Dividend size at the given \code{time} is
then expected to be equal to \code{fixed + proportional * S / S0}}
  }}
}
\value{
A list, with at least one entry for \code{price} resulting from
 \code{find_present_value()}.  For each greek requested, the list will have another entry.
}
\description{
Price the given instruments using \code{\link{find_present_value}}, and also compute the specified greeks
via finite-difference calculations.  The finite differences are higher-order
2-sided estimates where feasible, otherwise they are 1-sided estimates.  Bump
sizes are automatically chosen by \code{\link{resolve_bumps}}, unless the corresponding
\code{dS}, \code{dr}, \code{dvola} or \code{dhazard} override has been given a
non-\code{NULL} value.
}
\details{
If interest rates, default intensity, or volatility is zero then the corresponding
greek will \emph{not} be computed without a specifically-chosen user override
for \code{dr}, \code{dvola} or \code{dhazard}.

Bump sizes ultimately used, and central/one-sided estimate status, are available as as \code{attributes()} of the result
}
\examples{
S0=241.80
varc = variance_cumulation_from_vols(
  data.frame(
    time=c(0.13, 0.38, 0.63, 0.72, 1.72),
    volatility=c(0.48, 0.46, 0.46, 0.45, 0.45)
  ))
disct_fcn = spot_to_df_fcn(
  data.frame(time=c(1, 5, 10, 15),rate=c(0.01, 0.02, 0.03, 0.05))
)
cb = ConvertibleBond(
maturity=2.87, conversion_ratio=2.7788, notional=1000,
coupons=data.frame(
payment_time=seq(2.8, 0, by=-0.25),
    payment_size=1000*0.0025/4),
  discount_factor_fcn = disct_fcn
)
price_and_greeks = find_greeks(
  greeks=c("delta", "vega", "credit_dv01"),
  S0=S0,
instruments=list(CB=cb),
  num_time_steps=55,
  default_intensity_fcn = function(t, S, ...){0.03 + 0.01 * (S0/S)^1.5},
  discount_factor_fcn = disct_fcn,
  variance_cumulation_fcn=varc)
round(unlist(price_and_greeks),4)
}
\seealso{
Other Greeks:
\code{\link{GREEK_NAMES}},
\code{\link[=construct_descending_bumps]{construct_descending_bumps()}},
\code{\link[=greek_by_fd]{greek_by_fd()}},
\code{\link[=grid_delta_gamma]{grid_delta_gamma()}},
\code{\link[=resolve_bumps]{resolve_bumps()}},
\code{\link[=robust_greek]{robust_greek()}},
\code{\link[=safe_reprice]{safe_reprice()}}
}
\concept{Greeks}
