% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeks.R
\name{GREEK_NAMES}
\alias{GREEK_NAMES}
\title{Greeks available for computation, presently  "delta", "gamma", "theta", "vega",
"rates_dv01", and "credit_dv01".  Note in particular that we use the term
\code{rates_dv01} rather than \code{rho}}
\usage{
GREEK_NAMES
}
\description{
Greeks available for computation, presently  "delta", "gamma", "theta", "vega",
"rates_dv01", and "credit_dv01".  Note in particular that we use the term
\code{rates_dv01} rather than \code{rho}
}
\seealso{
Other Greeks:
\code{\link[=construct_descending_bumps]{construct_descending_bumps()}},
\code{\link[=find_greeks]{find_greeks()}},
\code{\link[=greek_by_fd]{greek_by_fd()}},
\code{\link[=grid_delta_gamma]{grid_delta_gamma()}},
\code{\link[=resolve_bumps]{resolve_bumps()}},
\code{\link[=robust_greek]{robust_greek()}},
\code{\link[=safe_reprice]{safe_reprice()}}
}
\concept{Greeks}
