% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit.R
\name{infer_conforming_time_grid}
\alias{infer_conforming_time_grid}
\title{A time grid with extra times inserted for coupons, calls and puts}
\usage{
infer_conforming_time_grid(min_num_time_steps, Tmax, instruments = NULL)
}
\arguments{
\item{min_num_time_steps}{The minimum number of timesteps the output vector should have}

\item{Tmax}{The maximum time on the grid}

\item{instruments}{A set of instruments whose maturity and terms
and conditions can introduce extra timesteps.  Each will be queried for the output of
a \code{critical_times} function.}
}
\description{
At its base, this function chooses a time grid with \code{1+min_num_time_steps}
  elements from 0 to \code{Tmax}.  Any coupon, call, or put times occurring in
  one of the supplied instruments are also inserted.
}

