% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_code.R
\name{is.blank}
\alias{is.blank}
\title{Return \code{TRUE} if the argument is empty, \code{NULL} or \code{NA}}
\usage{
is.blank(x, false.triggers = FALSE)
}
\arguments{
\item{x}{Argument to test}

\item{false.triggers}{Whether to allow nonempty vectors of all \code{FALSE} to trigger this condition}
}
\description{
Return \code{TRUE} if the argument is empty, \code{NULL} or \code{NA}
}
