% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackscholes.R
\docType{data}
\name{PUT}
\alias{PUT}
\title{Constant PUT for defining option contracts}
\format{An object of class \code{numeric} of length 1.}
\usage{
PUT
}
\description{
Constant PUT for defining option contracts
}
\keyword{datasets}
