% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_structures.R
\name{Quandl_df_fcn_UST_raw}
\alias{Quandl_df_fcn_UST_raw}
\title{Get a US Treasury curve discount factor function}
\usage{
Quandl_df_fcn_UST_raw(on_date)
}
\arguments{
\item{on_date}{Date for which to query Quandl for the curve}
}
\value{
A function taking two time arguments, which returns the discount factor from the second to the first
}
\description{
Get a US Treasury curve discount factor function
}
