% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ragtop-package}
\alias{ragtop-package}
\title{ragtop: Pricing Equity Derivatives with Extensions of Black-Scholes}
\description{
Algorithms to price American and European equity options, convertible bonds and a variety of other financial derivatives. It uses an extension of the usual Black-Scholes model in which jump to default may occur at a probability specified by a power-law link between stock price and hazard rate as found in the paper by Takahashi, Kobayashi, and Nakagawa (2001) \doi{10.3905/jfi.2001.319302}. We use ideas and techniques from Andersen and Buffum (2002) \doi{10.2139/ssrn.355308} and Linetsky (2006) \doi{10.1111/j.1467-9965.2006.00271.x}.
}
\author{
\strong{Maintainer}: Brian K. Boonstra \email{ragtop@boonstra.org}

}
\keyword{internal}
