% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeks.R
\name{safe_reprice}
\alias{safe_reprice}
\title{Reprice instruments with a bumped parameter}
\usage{
safe_reprice(reprice_runner, bump, instrument_names, greek = "Unspecified")
}
\arguments{
\item{reprice_runner}{A function that reprices instruments, possibly unsafely (for example if a bump results in negative variance)}

\item{bump}{The change in model parameter to be applied}

\item{instrument_names}{The names of instruments being analyzed}

\item{greek}{The name of the greek being bumped}
}
\value{
Instrument prices, with the given names
}
\description{
Evaluate one repricing of the whole instrument set, catching all exceptions, with NA wherever a
price is missing or non-finite.  A thrown error fails every instrument at once.
}
\seealso{
Other Greeks:
\code{\link{GREEK_NAMES}},
\code{\link[=construct_descending_bumps]{construct_descending_bumps()}},
\code{\link[=find_greeks]{find_greeks()}},
\code{\link[=greek_by_fd]{greek_by_fd()}},
\code{\link[=grid_delta_gamma]{grid_delta_gamma()}},
\code{\link[=resolve_bumps]{resolve_bumps()}},
\code{\link[=robust_greek]{robust_greek()}}
}
\concept{Greeks}
\keyword{internal}
