% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflows.R
\name{shift_for_dividends}
\alias{shift_for_dividends}
\title{Shift a set of grid values for dividends paid, using spline interpolation}
\usage{
shift_for_dividends(grid_values_before_shift, stock_prices, div_sum)
}
\description{
Shift a set of grid values for dividends paid, using spline interpolation
}

