% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{TIME_RESOLUTION_FACTOR}
\alias{TIME_RESOLUTION_FACTOR}
\title{Constant to define when times are considered so close to each other
 that they should be treated as simultaneous}
\usage{
TIME_RESOLUTION_FACTOR
}
\description{
Constant to define when times are considered so close to each other
 that they should be treated as simultaneous
}
