% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{TIME_RESOLUTION_SIGNIF_DIGITS}
\alias{TIME_RESOLUTION_SIGNIF_DIGITS}
\title{Constant to define when times are considered so close to each other
 that they should be treated as simultaneous, in terms of significant
 digits}
\usage{
TIME_RESOLUTION_SIGNIF_DIGITS
}
\description{
Constant to define when times are considered so close to each other
 that they should be treated as simultaneous, in terms of significant
 digits
}
