% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_structures.R
\name{treasury_df_raw}
\alias{treasury_df_raw}
\title{Get a US Treasury curve discount factor function}
\usage{
treasury_df_raw(on_date)
}
\arguments{
\item{on_date}{Date for which to query for the curve, year-month-day format}
}
\value{
A function taking two time arguments, which returns the discount factor from the second to the first
}
\description{
Get a US Treasury curve discount factor function
}
