% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSLAMarket}
\alias{TSLAMarket}
\title{Market information snapshot for TSLA options}
\format{
A list with these prices and rates
}
\usage{
data(TSLAMarket)
}
\description{
A dataset containing option contract details and a
 snapshot of market prices
 for Tesla Motors (TSLA) equity options, interest rates and an equity price.
}
\details{
The TSLAMarket list contains three elements:
\describe{
\item{\code{S0}}{The stock price as of snapshot time}
\item{\code{risk_free_rates}}{The spot risk-free rate curve as of snapshot time}
\item{\code{options}}{A data frame with details of the options market}
}
}
\keyword{datasets}
