% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\docType{class}
\name{ZeroCouponBond-class}
\alias{ZeroCouponBond-class}
\alias{ZeroCouponBond}
\title{A simple contract paying the \code{notional} amount at the \code{maturity}}
\description{
A simple contract paying the \code{notional} amount at the \code{maturity}
}
\section{Fields}{

\describe{
\item{\code{notional}}{The amount that will be paid at \code{maturity}, conditional on survival}

\item{\code{recovery_rate}}{The proportion of notional that would be expected to be paid to bond holders after bankruptcy court proceedings}

\item{\code{discount_factor_fcn}}{A function specifying how cashflows should generally be discounted for this instrument}
}}

\section{Methods}{

\describe{
\item{\code{optionality_fcn(v, ...)}}{Return a version of {v} at time {t} corrected for any optionality conditions.}

\item{\code{recovery_fcn(v, S, t, ...)}}{Return recovery value, given non-default values {v} at time {t}.  Subclasses may be more elaborate, this method simply returns 0.0.}
}}

