% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_code.R
\name{is.blank}
\alias{is.blank}
\title{Return TRUE if the argument is empty, NULL or NA}
\usage{
is.blank(x, false.triggers = FALSE)
}
\arguments{
\item{x}{Argument to test}

\item{false.triggers}{Whether to allow nonempty vectors of all FALSE to trigger this condition}
}
\description{
Return TRUE if the argument is empty, NULL or NA
}
