% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_structures.R
\name{Quandl_df_fcn_UST}
\alias{Quandl_df_fcn_UST}
\title{Get a US Treasury curve discount factor function}
\usage{
Quandl_df_fcn_UST(..., envir = parent.frame())
}
\arguments{
\item{...}{Arguments passed to \code{\link{Quandl_df_fcn_UST_raw}}}

\item{envir}{Environment passed to \code{\link{Quandl_df_fcn_UST_raw}}}
}
\value{
A function taking two time arguments, which returns the discount factor from the second to the first
}
\description{
This is a caching wrapper for \code{\link{Quandl_df_fcn_UST_raw}}
}

