% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{TIME_RESOLUTION_FACTOR}
\alias{TIME_RESOLUTION_FACTOR}
\title{Constant to define when times are considered so close to each other
 that they should be treated as simultaneous}
\format{An object of class \code{numeric} of length 1.}
\usage{
TIME_RESOLUTION_FACTOR
}
\description{
Constant to define when times are considered so close to each other
 that they should be treated as simultaneous
}
\keyword{datasets}

