% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSLAMarket}
\alias{TSLAMarket}
\title{Market information snapshot for TSLA options}
\format{An object of class \code{list} of length 3.}
\usage{
TSLAMarket
}
\description{
A dataset containing option contract details and a
 snapshot of market prices
 for Tesla Motors (TSLA) equity options.
}
\keyword{datasets}

